StickerApp/requirements.txt
pandas
openpyxl
pillow
reportlab
pdf2image
svglib
tk